network module: update tests for generating dracut arguments
