+++ tests/nosetests/pyanaconda_tests/network_test.py
@@ -289,128 +289,3 @@ class NetworkKSDataTests(unittest.TestCase):
         # network --hostname newhostname
         self.assertEqual(ksdata.network.network[0].hostname, "")
         self.assertEqual(ksdata.network.network[1].hostname, "newhostname")
-
-class NetworkIfcfgTests(unittest.TestCase):
-    def ifcfg_mock(self, settings):
-        def get(value):
-            return settings.get(value.upper(), "")
-        m = mock.Mock()
-        m.get = get
-        return m
-
-    def dracutBootArguments_test(self):
-        ifcfg = self.ifcfg_mock({"BOOTPROTO": "ibft"})
-        self.assertEqual(
-                network.dracutBootArguments("em1", ifcfg, ""),
-                set(["rd.iscsi.ibft"]))
-
-        ifcfg = self.ifcfg_mock({"BOOTPROTO": "ibft",
-                                 "HWADDR": "00:00:00:00:00:00"})
-        self.assertEqual(
-                network.dracutBootArguments("em1", ifcfg, ""),
-                set(["rd.iscsi.ibft"]))
-
-        ifcfg = self.ifcfg_mock({"BOOTPROTO": "dhcp"})
-        self.assertEqual(
-                network.dracutBootArguments("em1", ifcfg, "10.34.102.77"),
-                set(["ip=em1:dhcp"]))
-
-        ifcfg = self.ifcfg_mock({"BOOTPROTO": "dhcp",
-                                 "HWADDR": "00:00:00:00:00:00"})
-        self.assertEqual(
-                network.dracutBootArguments("em1", ifcfg, "10.34.102.77"),
-                set(["ip=em1:dhcp", "ifname=em1:00:00:00:00:00:00"]))
-
-        ifcfg = self.ifcfg_mock({"BOOTPROTO": "static",
-                                 "IPADDR": "10.34.102.233",
-                                 "GATEWAY": "10.34.102.254",
-                                 "NETMASK": "255.255.255.0",
-                                 "HWADDR": "00:00:00:00:00:00"})
-        self.assertEqual(
-                network.dracutBootArguments("em1", ifcfg, "10.34.102.77"),
-                set(["ip=10.34.102.233::10.34.102.254:255.255.255.0::em1:none", "ifname=em1:00:00:00:00:00:00"]))
-
-        ifcfg = self.ifcfg_mock({"BOOTPROTO": "static",
-                                 "IPADDR": "10.34.102.233",
-                                 "GATEWAY": "10.34.102.254",
-                                 "NETMASK": "255.255.255.0"})
-        self.assertEqual(
-                network.dracutBootArguments("em1", ifcfg, "10.34.102.77"),
-                set(["ip=10.34.102.233::10.34.102.254:255.255.255.0::em1:none"]))
-
-        ifcfg = self.ifcfg_mock({"BOOTPROTO": "static",
-                                 "IPADDR": "10.34.102.233",
-                                 "GATEWAY": "10.34.102.254",
-                                 "PREFIX": "24"})
-        self.assertEqual(
-                network.dracutBootArguments("em1", ifcfg, "10.34.102.77"),
-                set(["ip=10.34.102.233::10.34.102.254:255.255.255.0::em1:none"]))
-
-        ifcfg = self.ifcfg_mock({"BOOTPROTO": "static",
-                                 "IPADDR": "10.34.102.233",
-                                 "NETMASK": "255.255.255.0"})
-        self.assertEqual(
-                network.dracutBootArguments("em1", ifcfg, "10.34.102.77"),
-                set(["ip=10.34.102.233:::255.255.255.0::em1:none"]))
-
-        ifcfg = self.ifcfg_mock({"BOOTPROTO": "static",
-                                 "IPADDR": "10.34.102.233",
-                                 "GATEWAY": "10.34.102.254",
-                                 "NETMASK": "255.255.255.0"})
-        self.assertEqual(
-                network.dracutBootArguments("em1", ifcfg, "10.34.102.77", hostname="node1"),
-                set(["ip=10.34.102.233::10.34.102.254:255.255.255.0:node1:em1:none"]))
-
-        ifcfg = self.ifcfg_mock({"DHCPV6C": "yes",
-                                 "IPV6_AUTOCONF": "yes"})
-        self.assertEqual(
-                network.dracutBootArguments("em1", ifcfg, "2001::1"),
-                set(["ip=em1:dhcp6"]))
-
-        ifcfg = self.ifcfg_mock({"IPV6_AUTOCONF": "yes"})
-        self.assertEqual(
-                network.dracutBootArguments("em1", ifcfg, "2001::1"),
-                set(["ip=em1:auto6"]))
-
-        ifcfg = self.ifcfg_mock({"IPV6_AUTOCONF": "yes",
-                                 "HWADDR": "00:00:00:00:00:00"})
-        self.assertEqual(
-                network.dracutBootArguments("em1", ifcfg, "2001::1"),
-                set(["ip=em1:auto6", "ifname=em1:00:00:00:00:00:00"]))
-
-        ifcfg = self.ifcfg_mock({"IPV6ADDR": "2001::4"})
-        self.assertEqual(
-                network.dracutBootArguments("em1", ifcfg, "2001::1"),
-                set(["ip=[2001::4]:::::em1:none"]))
-
-        ifcfg = self.ifcfg_mock({"IPV6ADDR": "2001::4"})
-        self.assertEqual(
-                network.dracutBootArguments("em1", ifcfg, "2001::1", hostname="node1"),
-                set(["ip=[2001::4]::::node1:em1:none"]))
-
-        ifcfg = self.ifcfg_mock({"IPV6ADDR": "2001::4",
-                                 "IPV6_DEFAULTGW": "2001::a"})
-        self.assertEqual(
-                network.dracutBootArguments("em1", ifcfg, "2001::1"),
-                set(["ip=[2001::4]::[2001::a]:::em1:none"]))
-
-    @patch("blivet.arch.is_s390", lambda: True)
-    def dracutBootArguments_s390_test(self):
-        ifcfg = self.ifcfg_mock({"NETTYPE": "qeth",
-                                 "SUBCHANNELS": "0.0.f5f0,0.0.f5f1,0.0.f5f2",
-                                 "OPTIONS": '"layer2=1 portname=OSAPORT"'})
-        self.assertEqual(
-                network.dracutBootArguments("eth0", ifcfg, ""),
-                set(["rd.znet=qeth,0.0.f5f0,0.0.f5f1,0.0.f5f2,layer2=1,portname=OSAPORT"]))
-
-        ifcfg = self.ifcfg_mock({"NETTYPE": "qeth",
-                                 "SUBCHANNELS": "0.0.f5f0,0.0.f5f1,0.0.f5f2",
-                                 "OPTIONS": '"layer2=1 portname=OSAPORT"',
-                                 "BOOTPROTO": "static",
-                                 "IPADDR": "10.34.102.233",
-                                 "GATEWAY": "10.34.102.254",
-                                 "NETMASK": "255.255.255.0"})
-        self.assertEqual(
-                network.dracutBootArguments("eth0", ifcfg, "10.34.102.77"),
-                set(["rd.znet=qeth,0.0.f5f0,0.0.f5f1,0.0.f5f2,layer2=1,portname=OSAPORT",
-                     "ip=10.34.102.233::10.34.102.254:255.255.255.0::eth0:none"]))
